requirements.txt

openai>=1.0.0
requests
playwright
playwright-stealth
python-dotenv
PyYAML>=6.0
pytest>=7.0
